+++ trimesh/exchange/dae.py
